use pre_fit in OMP
